requirements.txt
streamlit
pandas
numpy
matplotlib
PyDeck
seaborn
plotly-express

